Refactor unit tests for SharesNavigation component
