+++ packages/web-app-files/src/components/AppBar/SharesNavigation.vue
@@ -41,7 +41,7 @@
   </nav>
 </template>
 
-<script>
+<script lang="ts">
 import {
   isLocationSharesActive,
   locationSharesViaLink,
